Rearrange things in the UK109 EPROMs.
Move the EQU directives so that they all come before the data and code. Move a few routines so that they're still close to 'vduchar', but not within that routine. Add a few more indirect jumps in the table at the end of the ROM. Initialise the saved registers in RAM (to zero) so that they don't come out as random garbage in the 'R' command.