Switch over to using the 'asm6809' assembler.
Use 'sed' to convert the source code for use with the old Frankensein Assembler, in M6809 mode only. Use real HD6309 mnemonics now. Simplify 'if' commands to a form that both assemblers will accept. Add 'setdp' directive for 'asm6809' and comment it out for the old assembler. Use HD6309 'tfm' for VDU scrolling.