Fix global declarations.
Global variables need to be defined in just one place in the C files and declared with 'extern' in the header files. Header files should not contain code that allocates storage. Not sure when this all changed, but it's at least 14 years since I last compiled this code. Hey, it even ran on the DEC Alpha at one time.